+++ pycroft/model/traffic.py
